Added coinflip command.
